Start on serialization work and get rid of RangeFinder
