+++ mp_slam/scanmatching.py
@@ -1,6 +1,8 @@
 from .__init__ import print_config, default_config
 
-from mp_slam_cpp import Wrapper, Pose2, ScanMatcherConfig
+from mp_slam.cpp import Pose2
+
+from mp_slam_cpp import Wrapper, ScanMatcherConfig
 from uuid import uuid4
 from tiny_tf.tf import Transform
